requirements.txt
absl-py==1.4.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
fsspec>=2023.5.0,<2026.0.0
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
packaging==25.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
setuptools==80.9.0
sympy==1.14.0
tokenizers==0.21.2
torch==2.4.1
tqdm==4.67.1
transformers==4.53.2
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
gunicorn
kaggle==1.6.17
datasets==3.0.2
accelerate>=0.32.1
peft==0.13.2
pandas==2.2.3
scikit-learn==1.5.2
google-generativeai==0.8.3
torchvision==0.19.1
matplotlib==3.9.2
